requirements.txt
pdf2docx==0.5.8
python-docx==0.8.11
PyPDF2==3.0.1
reportlab==4.0.4
docx2pdf==0.1.8
docx2txt==0.8
comtypes==1.4.0
